refactor: ARアセット作成ユースケース用にディレクトリ作成
